add a chance to crit and add type of characters in attack function

+++ JEUX PROJET 3 IOS DEV/Weapons/Weapon.swift
@@ -10,10 +10,10 @@ import Foundation
 
 class Weapon {
     
-    var damage: Int
-    var heal: Int
+    var damage: Decimal
+    var heal: Decimal
     
-    init(damage: Int, heal: Int) {
+    init(damage: Decimal, heal: Decimal) {
         self.damage = damage
         self.heal = heal
     }
